pos_system/requirements.txt
requests
qrcode
python-dotenv
Pillow
